Introduce @turnkey/encoding to consolidate utils across packages
