+++ .changeset/config.json
@@ -6,5 +6,5 @@
   "access": "public",
   "baseBranch": "main",
   "bumpVersionsWithWorkspaceProtocolOnly": true,
-  "ignore": ["@turnkey/jest-config", "@turnkey/example-*"]
+  "ignore": ["@turnkey/jest-config", "@turnkey/encoding", "@turnkey/example-*"]
 }
