+++ internal/encoding/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@turnkey/encoding",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "clean": "rimraf ./dist ./.cache",
+    "test": "jest",
+    "typecheck": "tsc -p tsconfig.typecheck.json"
+  },
+  "engines": {
+    "node": ">=18.0.0"
+  }
+}
