+++ internal/encoding/src/__tests__/encoding-test.ts
