+++ packages/http/package.json
@@ -48,6 +48,7 @@
   "dependencies": {
     "cross-fetch": "^3.1.5",
     "@turnkey/api-key-stamper": "workspace:*",
+    "@turnkey/encoding": "workspace:*",
     "@turnkey/webauthn-stamper": "workspace:*"
   },
   "engines": {
